tests/Document-createProcessingInstruction.html

<!DOCTYPE html>
<title>Document.createProcessingInstruction in HTML documents</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="../web-dom-core#dom-document-createprocessinginstruction">
<script src="testharness.js"></script>
<div id=log></div>
<script>
// maybe put this somewhere else?
test(function() {
  assert_throws("NOT_SUPPORTED_ERR", function() { document.createProcessingInstruction("A", "B") })
})
</script>
